Pong is Basically done
Can control the second paddle with the keyboard. I added controls directions to the pause menu.  The scores now update when one of the players score and it saves the scores (using player prefs). You can reset the scores with the button at the bottom of the screen. You can also now resize the window or maximize it, though it looks weird if it's not be the same dimensions.

Started to try to set up the webGL stuff.